DBZ-4169 Consolidate CacheProvider impls into Processor impls
