+++ debezium-connector-oracle/src/main/java/io/debezium/connector/oracle/logminer/processor/infinispan/RemoteInfinispanLogMinerEventProcessor.java
@@ -5,12 +5,29 @@
  */
 package io.debezium.connector.oracle.logminer.processor.infinispan;
 
-import java.util.Map;
+import static io.debezium.connector.oracle.OracleConnectorConfig.LOG_MINING_BUFFER_INFINISPAN_CACHE_COMMITTED_TRANSACTIONS;
+import static io.debezium.connector.oracle.OracleConnectorConfig.LOG_MINING_BUFFER_INFINISPAN_CACHE_EVENTS;
+import static io.debezium.connector.oracle.OracleConnectorConfig.LOG_MINING_BUFFER_INFINISPAN_CACHE_ROLLBACK_TRANSACTIONS;
+import static io.debezium.connector.oracle.OracleConnectorConfig.LOG_MINING_BUFFER_INFINISPAN_CACHE_SCHEMA_CHANGES;
+import static io.debezium.connector.oracle.OracleConnectorConfig.LOG_MINING_BUFFER_INFINISPAN_CACHE_TRANSACTIONS;
 
+import java.util.Map;
+import java.util.Objects;
+import java.util.Properties;
+
+import org.infinispan.client.hotrod.RemoteCache;
+import org.infinispan.client.hotrod.RemoteCacheManager;
+import org.infinispan.client.hotrod.configuration.Configuration;
+import org.infinispan.client.hotrod.configuration.ConfigurationBuilder;
+import org.infinispan.client.hotrod.impl.ConfigurationProperties;
+import org.infinispan.commons.api.BasicCache;
+import org.infinispan.commons.configuration.XMLStringConfiguration;
 import org.infinispan.commons.util.CloseableIterator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import io.debezium.DebeziumException;
+import io.debezium.config.Field;
 import io.debezium.connector.oracle.OracleConnection;
 import io.debezium.connector.oracle.OracleConnectorConfig;
 import io.debezium.connector.oracle.OracleDatabaseSchema;
@@ -18,17 +35,34 @@ import io.debezium.connector.oracle.OracleOffsetContext;
 import io.debezium.connector.oracle.OraclePartition;
 import io.debezium.connector.oracle.OracleStreamingChangeEventSourceMetrics;
 import io.debezium.connector.oracle.Scn;
+import io.debezium.connector.oracle.logminer.events.LogMinerEvent;
+import io.debezium.connector.oracle.logminer.processor.infinispan.marshalling.LogMinerEventMarshallerImpl;
+import io.debezium.connector.oracle.logminer.processor.infinispan.marshalling.TransactionMarshallerImpl;
 import io.debezium.pipeline.EventDispatcher;
 import io.debezium.pipeline.source.spi.ChangeEventSource.ChangeEventSourceContext;
 import io.debezium.relational.TableId;
 
 /**
+ * A concrete implementation of {@link AbstractInfinispanLogMinerEventProcessor} that uses Infinispan with
+ * the Hotrod client to store transaction and mined event data in caches.
+ *
+ * The cache configurations are supplied via connector configurations and are expected to be valid XML
+ * that represents parseable distributed cache setups for Infinispan.
+ *
  * @author Chris Cranford
  */
-public class RemoteInfinispanLogMinerEventProcessor extends AbstractInfinispanLogMinerEventProcessor {
+public class RemoteInfinispanLogMinerEventProcessor extends AbstractInfinispanLogMinerEventProcessor implements CacheProvider {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(RemoteInfinispanLogMinerEventProcessor.class);
-    private final RemoteCacheProvider cacheProvider;
+
+    private final RemoteCacheManager cacheManager;
+    private final boolean dropBufferOnStop;
+
+    private final RemoteCache<String, InfinispanTransaction> transactionCache;
+    private final RemoteCache<String, LogMinerEvent> eventCache;
+    private final RemoteCache<String, String> recentlyCommittedTransactionsCache;
+    private final RemoteCache<String, String> rollbackTransactionsCache;
+    private final RemoteCache<String, String> schemaChangesCache;
 
     public RemoteInfinispanLogMinerEventProcessor(ChangeEventSourceContext context,
                                                   OracleConnectorConfig connectorConfig,
@@ -40,29 +74,76 @@ public class RemoteInfinispanLogMinerEventProcessor extends AbstractInfinispanLo
                                                   OracleStreamingChangeEventSourceMetrics metrics) {
         super(context, connectorConfig, jdbcConnection, dispatcher, partition, offsetContext, schema, metrics);
 
-        cacheProvider = new RemoteCacheProvider(connectorConfig);
-        cacheProvider.displayCacheStatistics();
+        Configuration config = new ConfigurationBuilder()
+                .withProperties(getHotrodClientProperties(connectorConfig))
+                // todo: why must these be defined manually rather than automated like embedded mode?
+                .addContextInitializer(TransactionMarshallerImpl.class.getName())
+                .addContextInitializer(LogMinerEventMarshallerImpl.class.getName())
+                .build();
+
+        LOGGER.info("Using Infinispan in Hotrod client mode");
+        this.cacheManager = new RemoteCacheManager(config, true);
+        this.dropBufferOnStop = connectorConfig.isLogMiningBufferDropOnStop();
+
+        this.transactionCache = createCache(TRANSACTIONS_CACHE_NAME, connectorConfig, LOG_MINING_BUFFER_INFINISPAN_CACHE_TRANSACTIONS);
+        this.recentlyCommittedTransactionsCache = createCache(COMMIT_TRANSACTIONS_CACHE_NAME, connectorConfig, LOG_MINING_BUFFER_INFINISPAN_CACHE_COMMITTED_TRANSACTIONS);
+        this.rollbackTransactionsCache = createCache(ROLLBACK_TRANSACTIONS_CACHE_NAME, connectorConfig, LOG_MINING_BUFFER_INFINISPAN_CACHE_ROLLBACK_TRANSACTIONS);
+        this.schemaChangesCache = createCache(SCHEMA_CHANGES_CACHE_NAME, connectorConfig, LOG_MINING_BUFFER_INFINISPAN_CACHE_SCHEMA_CHANGES);
+        this.eventCache = createCache(EVENTS_CACHE_NAME, connectorConfig, LOG_MINING_BUFFER_INFINISPAN_CACHE_EVENTS);
+
+        displayCacheStatistics();
     }
 
     @Override
     public void close() throws Exception {
-        cacheProvider.close();
+        if (dropBufferOnStop) {
+            LOGGER.info("Clearing infinispan caches");
+            transactionCache.clear();
+            eventCache.clear();
+            schemaChangesCache.clear();
+            recentlyCommittedTransactionsCache.clear();
+            rollbackTransactionsCache.clear();
+
+            // this block should only be used by tests, should we wrap this in case admin rights aren't given?
+            cacheManager.administration().removeCache(CacheProvider.TRANSACTIONS_CACHE_NAME);
+            cacheManager.administration().removeCache(CacheProvider.COMMIT_TRANSACTIONS_CACHE_NAME);
+            cacheManager.administration().removeCache(CacheProvider.ROLLBACK_TRANSACTIONS_CACHE_NAME);
+            cacheManager.administration().removeCache(CacheProvider.SCHEMA_CHANGES_CACHE_NAME);
+            cacheManager.administration().removeCache(CacheProvider.EVENTS_CACHE_NAME);
+        }
+        LOGGER.info("Shutting down infinispan remote caches");
+        cacheManager.close();
     }
 
     @Override
-    protected Map<String, InfinispanTransaction> getTransactionCache() {
-        return cacheProvider.getTransactionCache();
+    public BasicCache<String, InfinispanTransaction> getTransactionCache() {
+        return transactionCache;
     }
 
     @Override
-    protected RemoteCacheProvider getCacheProvider() {
-        return cacheProvider;
+    public BasicCache<String, LogMinerEvent> getEventCache() {
+        return eventCache;
+    }
+
+    @Override
+    public BasicCache<String, String> getSchemaChangesCache() {
+        return schemaChangesCache;
+    }
+
+    @Override
+    public BasicCache<String, String> getCommittedTransactionsCache() {
+        return recentlyCommittedTransactionsCache;
+    }
+
+    @Override
+    public BasicCache<String, String> getRollbackTransactionsCache() {
+        return rollbackTransactionsCache;
     }
 
     @Override
     protected Scn getTransactionCacheMinimumScn() {
         Scn minimumScn = Scn.NULL;
-        try (CloseableIterator<InfinispanTransaction> iterator = cacheProvider.getTransactionCache().values().iterator()) {
+        try (CloseableIterator<InfinispanTransaction> iterator = transactionCache.values().iterator()) {
             while (iterator.hasNext()) {
                 final Scn transactionScn = iterator.next().getStartScn();
                 if (minimumScn.isNull()) {
@@ -75,7 +156,44 @@ public class RemoteInfinispanLogMinerEventProcessor extends AbstractInfinispanLo
                 }
             }
         }
-        LOGGER.info("Transaction Cache Min SCN {}", minimumScn);
         return minimumScn;
+    }
+
+    private Properties getHotrodClientProperties(OracleConnectorConfig connectorConfig) {
+        final Map<String, String> clientSettings = connectorConfig.getConfig()
+                .subset("log.mining.buffer.infinispan.client.", true)
+                .asMap();
+
+        final Properties properties = new Properties();
+        for (Map.Entry<String, String> entry : clientSettings.entrySet()) {
+            properties.put("infinispan.client." + entry.getKey(), entry.getValue());
+            if (entry.getKey().toLowerCase().endsWith(ConfigurationProperties.AUTH_USERNAME.toLowerCase())) {
+                // If an authentication username is supplied, enforce authentication required
+                properties.put(ConfigurationProperties.USE_AUTH, "true");
+            }
+        }
+        return properties;
+    }
+
+    private <C, V> RemoteCache<C, V> createCache(String cacheName, OracleConnectorConfig connectorConfig, Field field) {
+        Objects.requireNonNull(cacheName);
+
+        RemoteCache<C, V> cache = cacheManager.getCache(cacheName);
+        if (cache != null) {
+            // cache is already defined, simply return it
+            LOGGER.info("Remote cache '{}' already defined.", cacheName);
+            return cache;
+        }
+
+        final String cacheConfiguration = connectorConfig.getConfig().getString(field);
+        Objects.requireNonNull(cacheConfiguration);
+
+        cache = cacheManager.administration().createCache(cacheName, new XMLStringConfiguration(cacheConfiguration));
+        if (cache == null) {
+            throw new DebeziumException("Failed to create remote Infinispan cache: " + cacheName);
+        }
+
+        LOGGER.info("Created remote infinispan cache: {}", cacheName);
+        return cache;
     }
 }
